requirements.txt
pandas==2.3.0
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.8.6
streamlit==1.30.0
matplotlib==3.10.3
typing_extensions==4.15.0